change compiling and linking routine

+++ src/backend/tests/c-lisp/kernel/runtime/bench.c
@@ -2,10 +2,13 @@
 #include <stdlib.h>
 #include <time.h>
 
+// __attribute__((optnone))
 double flops(int m, int n, int k, double time) {
-    double num_ops = 2 * (m * n * k);
+    // printf("%.3f\n", time);      /** FOR DEBUGGING */
+    double num_ops = 2.0 * (double)m * (double)n * (double)k;
     return num_ops / (time * 1e9);
 }
+
 
 clock_t timer() {
     return clock();
